+++ frontend/packages/data-portal/app/components/BrowseData/BrowseDataSearch.tsx
@@ -1,10 +1,13 @@
 import { InputSearch } from '@czi-sds/components'
 import { useDebouncedEffect } from '@react-hookz/web'
 import { useSearchParams } from '@remix-run/react'
-import { useRef, useState } from 'react'
+import { useAtom } from 'jotai'
+import { useHydrateAtoms } from 'jotai/utils'
+import { useEffect, useRef } from 'react'
 
 import { QueryParams } from 'app/constants/query'
-import { i18n } from 'app/i18n'
+import { useI18n } from 'app/hooks/useI18n'
+import { searchQueryAtom } from 'app/state/search'
 
 /**
  * The amount of time to wait after the user has typed in a query before
@@ -15,7 +18,14 @@ const SEARCH_QUERY_DEBOUNCE_TIME_MS = 500
 
 export function BrowseDataSearch() {
   const [searchParams, setSearchParams] = useSearchParams()
-  const [query, setQuery] = useState(searchParams.get(QueryParams.Search) ?? '')
+  const [query, setQuery] = useAtom(searchQueryAtom)
+
+  useHydrateAtoms([
+    [searchQueryAtom, searchParams.get(QueryParams.Search) ?? ''],
+  ])
+
+  // Reset when navigating away
+  useEffect(() => () => setQuery(''), [setQuery])
 
   // If the user hasn't typed in a key for 500ms, then update the search params.
   const initialLoadRef = useRef(true)
@@ -41,14 +51,17 @@ export function BrowseDataSearch() {
     SEARCH_QUERY_DEBOUNCE_TIME_MS,
   )
 
+  const { t } = useI18n()
+
   return (
     <InputSearch
       id="data-search"
-      label={i18n.search}
-      sdsStyle="rounded"
-      placeholder={i18n.search}
+      label={t('search')}
+      sdsStyle="square"
+      placeholder={t('searchByDatasetName')}
       value={query}
       onChange={(event) => setQuery(event.target.value)}
+      className="!m-0 [&>.MuiInputBase-root]:!min-w-[240px]"
     />
   )
 }