+++ frontend/packages/data-portal/app/components/MLChallenge/MdxComponents/MdxContributorList.tsx
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react'
+
+import styles from './MdxContributorList.module.css'
+
+export function MdxContributorList({ children }: { children: ReactNode }) {
+  return <div className={styles.contributorList}>{children}</div>
+}