+++ frontend/packages/data-portal/app/components/MLChallenge/MdxComponents/MdxIconGrid.tsx
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react'
+
+import styles from './MdxIconGrid.module.css'
+
+export function MdxIconGrid({ children }: { children: ReactNode }) {
+  return <div className={styles.iconGrid}>{children}</div>
+}