+++ frontend/packages/data-portal/app/components/MLChallenge/MdxComponents/MdxTable.tsx
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react'
+
+import styles from './MdxTable.module.css'
+
+export function MdxTable({ children }: { children: ReactNode }) {
+  return <div className={styles.table}>{children}</div>
+}