+++ frontend/packages/data-portal/app/components/TablePageLayout/index.ts
@@ -1,0 +1,3 @@
+export * from './TableHeader'
+export * from './TableHeaderDefinition'
+export * from './TablePageLayout'